WIP: added web server and routes
