+++ config/config.dev.yml
@@ -6,3 +6,7 @@ redis:
 
 webserver:
   port: ':9001'
+  tpl: './resources/tpl/'
+  tpl_pattern: '*.html'
+  static_path: './resources/static/'
+  static_prefix: '/static/'
